# openapi/ChikExams.openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Chik.Exams",
    "description": "Chik.Exams API",
    "contact": {
      "name": "Chik.Exams",
      "email": "support@chik.ng"
    },
    "version": "1.0.1"
  },
  "paths": {
...
    "/api/quizzes": {
      "post": {
        "tags": ["Quizzes"],
        "operationId": "Quizzes_Create",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Api.CreateQuizRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Api.CreateQuizRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Api.CreateQuizRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Api.CreateQuizRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Quiz"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quiz"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quiz"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Quizzes"],
        "operationId": "Quizzes_Search",
        "parameters": [
          {
            "name": "Title",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "CreatorId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "ExaminerId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "DateRange.From",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "DateRange.To",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "QuizIds",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          },
          {
            "name": "IncludeCreator",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "IncludeExaminer",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "IncludeQuestions",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Rows",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Quiz"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Quiz"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Quiz"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/quizzes/{id}": {
      "get": {
        "tags": ["Quizzes"],
        "operationId": "Quizzes_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "includeQuestions",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Quiz"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quiz"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quiz"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Quizzes"],
        "operationId": "Quizzes_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Api.UpdateQuizRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Api.UpdateQuizRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Api.UpdateQuizRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Api.UpdateQuizRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Quiz"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quiz"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quiz"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Quizzes"],
        "operationId": "Quizzes_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/quizzes/{id}/export": {
      "get": {
        "tags": ["Quizzes"],
        "operationId": "Quizzes_Export",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/quizzes/import": {
      "post": {
        "tags": ["Quizzes"],
        "operationId": "Quizzes_Import",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "examinerId": {
                    "type": "integer",
                    "format": "int64"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                },
                "examinerId": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Quiz"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quiz"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quiz"
                }
              }
            }
          }
        }
      }
    },
    "/api/quizzes/{quizId}/questions": {
      "get": {
        "tags": ["Quizzes"],
        "operationId": "Quizzes_GetQuestions",
        "parameters": [
          {
            "name": "quizId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "includeDeactivated",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuizQuestion"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuizQuestion"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuizQuestion"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Quizzes"],
        "operationId": "Quizzes_CreateQuestion",
        "parameters": [
          {
            "name": "quizId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Api.CreateQuizQuestionRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Api.CreateQuizQuestionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Api.CreateQuizQuestionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Api.CreateQuizQuestionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/QuizQuestion"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizQuestion"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizQuestion"
                }
              }
            }
          }
        }
      }
    },
    "/api/quizzes/{quizId}/questions/reorder": {
      "post": {
        "tags": ["Quizzes"],
        "operationId": "Quizzes_ReorderQuestions",
        "parameters": [
          {
            "name": "quizId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Api.ReorderQuestionsRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Api.ReorderQuestionsRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Api.ReorderQuestionsRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Api.ReorderQuestionsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Api.CreateQuizQuestionRequest": {
        "required": ["order", "prompt", "properties", "score", "typeId"],
        "type": "object",
        "properties": {
          "prompt": {
            "type": "string",
            "nullable": true
          },
          "typeId": {
            "type": "integer",
            "format": "int64"
          },
          "properties": {
            "type": "string",
            "nullable": true
          },
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "order": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Api.CreateQuizRequest": {
        "required": ["description", "title"],
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "examinerId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "duration": {
            "type": "string",
            "format": "date-span",
            "nullable": true
          }
        }
      },
...
      "Api.ReorderQuestionsRequest": {
        "required": ["questionIdsInOrder"],
        "type": "object",
        "properties": {
          "questionIdsInOrder": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "nullable": true
          }
        }
      },
...
      "Api.UpdateQuizRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "examinerId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "duration": {
            "type": "string",
            "format": "date-span",
            "nullable": true
          }
        }
      },
...
      "Quiz": {
        "required": ["createdAt", "creatorId", "description", "id", "title"],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "creatorId": {
            "type": "integer",
            "format": "int64"
          },
          "examinerId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "duration": {
            "type": "string",
            "format": "date-span",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "creator": {
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ],
            "nullable": true
          },
          "examiner": {
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ],
            "nullable": true
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuizQuestion"
            },
            "nullable": true
          }
        }
      },
      "QuizQuestion": {
        "required": [
          "createdAt",
          "id",
          "isActive",
          "order",
          "prompt",
          "quizId",
          "score",
          "typeId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "quizId": {
            "type": "integer",
            "format": "int64"
          },
          "prompt": {
            "type": "string",
            "nullable": true
          },
          "typeId": {
            "type": "integer",
            "format": "int32"
          },
          "properties": {
            "allOf": [
              {
                "$ref": "#/components/schemas/QuizQuestion.QuestionType"
              }
            ],
            "nullable": true
          },
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deactivatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "quiz": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Quiz"
              }
            ],
            "nullable": true
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/QuizQuestionType"
              }
            ],
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "readOnly": true
          }
        }
      },
      "QuizQuestion.QuestionType": {
        "required": ["type"],
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
      "QuizQuestionType": {
        "required": ["createdAt", "description", "id", "name"],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuizQuestion"
            },
            "nullable": true
          }
        }
      },
      "User": {
        "required": ["createdAt", "id", "roles", "username"],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserRole"
            },
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastLogin": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "UserRole": {
        "enum": [1, 2, 4],
        "type": "integer",
        "description": "",
        "format": "int32",
        "x-enum-varnames": ["Admin", "Teacher", "Student"]
      }
    }
  }
}